app/pyproject.toml
[tool.poetry]
name = "biquery"
version = "0.1.0"
description = ""
authors = ["MrTrustworthy <tinywritingmakesmailhardtoread@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-bigquery = "^1.21"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
